Added features/screens folders — created Search bar features and restaurant (list/results) screens — Feature branch *new*
